+++ core/src/main/kotlin/net/corda/core/transactions/TransactionWithSignatures.kt
@@ -8,6 +8,7 @@ import net.corda.core.utilities.toNonEmptySet
 import java.security.InvalidKeyException
 import java.security.PublicKey
 import java.security.SignatureException
+import java.util.*
 
 /** An interface for transactions containing signatures, with logic for signature verification */
 interface TransactionWithSignatures : NamedByHash {
@@ -24,7 +25,7 @@ interface TransactionWithSignatures : NamedByHash {
      * @throws SignaturesMissingException if any signatures should have been present but were not.
      */
     @Throws(SignatureException::class)
-    fun verifyRequiredSignatures() = verifySignaturesExcept()
+    fun verifyRequiredSignatures() = verifySignaturesExcept(emptySet())
 
     /**
      * Verifies the signatures on this transaction and throws if any are missing which aren't passed as parameters.
@@ -40,6 +41,23 @@ interface TransactionWithSignatures : NamedByHash {
      */
     @Throws(SignatureException::class)
     fun verifySignaturesExcept(vararg allowedToBeMissing: PublicKey) {
+        verifySignaturesExcept(Arrays.asList(*allowedToBeMissing))
+    }
+
+    /**
+     * Verifies the signatures on this transaction and throws if any are missing which aren't specified to be missing.
+     * In this context, "verifying" means checking they are valid signatures and that their public keys are in
+     * the [requiredSigningKeys] set.
+     *
+     * Normally you would not provide any keys to this function, but if you're in the process of building a partial
+     * transaction and you want to access the contents before you've signed it, you can specify your own keys here
+     * to bypass that check.
+     *
+     * @throws SignatureException if any signatures are invalid or unrecognised.
+     * @throws SignaturesMissingException if any signatures should have been present but were not.
+     */
+    @Throws(SignatureException::class)
+    fun verifySignaturesExcept(allowedToBeMissing: Collection<PublicKey>) {
         checkSignaturesAreValid()
 
         val needed = getMissingSignatures() - allowedToBeMissing
@@ -51,7 +69,7 @@ interface TransactionWithSignatures : NamedByHash {
      * Mathematically validates the signatures that are present on this transaction. This does not imply that
      * the signatures are by the right keys, or that there are sufficient signatures, just that they aren't
      * corrupt. If you use this function directly you'll need to do the other checks yourself. Probably you
-     * want [verifySignatures] instead.
+     * want [verifyRequiredSignatures] instead.
      *
      * @throws InvalidKeyException if the key on a signature is invalid.
      * @throws SignatureException if a signature fails to verify.
@@ -75,7 +93,6 @@ interface TransactionWithSignatures : NamedByHash {
         val sigKeys = sigs.map { it.by }.toSet()
         // TODO Problem is that we can get single PublicKey wrapped as CompositeKey in allowedToBeMissing/mustSign
         //  equals on CompositeKey won't catch this case (do we want to single PublicKey be equal to the same key wrapped in CompositeKey with threshold 1?)
-        val missing = requiredSigningKeys.filter { !it.isFulfilledBy(sigKeys) }.toSet()
-        return missing
+        return requiredSigningKeys.filter { !it.isFulfilledBy(sigKeys) }.toSet()
     }
 }